modified:   src/main/java/com/plagarism/detect/co- ntrollers/ScraperController.java deleted:    src/main/java/com/plagarism/detect/reader/App.java modified:   src/main/java/com/plagarism/detect/reader/DocumentReader.java final changes
