+++ src/main/java/com/plagarism/detect/controllers/ScraperController.java
@@ -7,12 +7,20 @@ import com.plagarism.detect.domain.Queries;
 import com.plagarism.detect.domain.QueriesDTO;
 import com.plagarism.detect.domain.Query;
 import com.plagarism.detect.domain.QueryDTO;
+import com.plagarism.detect.reader.DocumentReader;
+
+import java.io.File;
+import java.util.ArrayList;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
 @RestController
 public class ScraperController {
+
+   // NOTE this needs to be changed once deployed
+   public static final String FILE_PATH = "C:\\Users\\smhal\\Documents\\Coding\\java\\docReader\\demo\\src\\main\\java\\com\\docs\\";
+   
    // TODO reposiories if needed
 
    /*
@@ -29,9 +37,10 @@ public class ScraperController {
     * 7. Error catcher
     */
 
-    /* TODO
-     * getQueryResults
-     */
+   /*
+    * TODO
+    * getQueryResults
+    */
    @GetMapping(value = "/scraper")
    public Queries getQueryResults(@RequestBody QueriesDTO queries) {
       Queries newQueries = queriesDTOtoQueries(queries);
@@ -64,7 +73,8 @@ public class ScraperController {
       return null;
    }
 
-   /* TODO
+   /*
+    * TODO
     * info send a text response of basic information of endpoints on the API
     */
    @GetMapping(value = "/info")
@@ -72,15 +82,18 @@ public class ScraperController {
       return "Info on API: (Under Construction)";
    }
 
-   /* TODO
-    * this is he home or landing page mostly as a test to make sure the API is running.
+   /*
+    * TODO
+    * this is he home or landing page mostly as a test to make sure the API is
+    * running.
     */
    @GetMapping(value = "/")
    public String home() {
       return "Plagarism Detection Service.";
    }
-   
-   /* TODO
+
+   /*
+    * TODO
     * This is a catch-all error page that gives a basic non-descript error report.
     */
    @GetMapping(value = "/error")
@@ -92,14 +105,14 @@ public class ScraperController {
     * Helper functions:
     * 1. scraper
     * 2. DTO converter
-    *    a. QueriesDTO -> Queries
-    *    b. QueryDTO -> Query
+    * a. QueriesDTO -> Queries
+    * b. QueryDTO -> Query
     */
    private Queries useScraper(Queries quereies) {
       return null;
    }
 
-   private Queries queriesDTOtoQueries(QueriesDTO quereiesDTO){
+   private Queries queriesDTOtoQueries(QueriesDTO quereiesDTO) {
       Queries queries = new Queries();
       for (QueryDTO queryDTO : quereiesDTO.getListOfQueries()) {
          Query newQuery = queryDTOtoQuery(queryDTO);
@@ -108,11 +121,61 @@ public class ScraperController {
       return queries;
    }
 
-   private Query queryDTOtoQuery(QueryDTO queryDTO){
+   private Query queryDTOtoQuery(QueryDTO queryDTO) {
       Query query = new Query();
       query.setQueryText(queryDTO.queryText);
       query.setFoundOnline(queryDTO.foundOnline);
       return query;
    }
 
+   private void testDocumentReader(){
+      DocumentReader reader = new DocumentReader();
+      ArrayList<String> fileLocations = findDocuments();
+
+      for (String fileLocation : fileLocations) {
+         System.out.println("Reading document @ " + fileLocation);
+         reader.setDocument(fileLocation);
+         // make sure I am reading the documents properly test
+         // reader.readDocument();
+         try {
+            reader.findQuestions();
+         } catch (Exception e) {
+
+            System.out.println("error finding questions error = " + e);
+         }
+         // now we read out questions found if any
+         // ArrayList<String> questions = reader.getQuestions();
+         // if (reader.getQuestions().size() == 0) {
+         // System.out.println("No questions found");
+         // } else {
+         // System.out.println("Questions found");
+         // int numberQuestion = 1;
+         // for (String question : questions) {
+         // System.out.println("" + (numberQuestion++) + " - " + question);
+         // }
+         // }
+      }
+   }
+   
+   private static ArrayList<String> findDocuments() {
+      // Folder path and initialization of returned list
+      File folder = new File(FILE_PATH);
+      ArrayList<String> listOfFileLocations = new ArrayList<String>();
+
+      // grab all the files
+      File[] listOfFiles = folder.listFiles();
+
+      // grab all the files and dir at the location, then put them into the list
+      for (int i = 0; i < listOfFiles.length; i++) {
+          if (listOfFiles[i].isFile()) {
+              // System.out.println(listOfFiles[i].getName()); // check to make sure all files
+              // are grabbed
+
+              // Need to format the file location to get all exact paths
+              listOfFileLocations.add(FILE_PATH + listOfFiles[i].getName());
+          }
+      }
+
+      return listOfFileLocations;
+  }
 }
